feat: create birth component
